requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asttokens==2.2.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
discord.py==2.3.2
executing==1.2.0
ffmpeg-python==0.2.0
fonttools==4.39.3
frozenlist==1.3.3
future==0.18.3
idna==3.4
imageio==2.28.0
imutils==0.5.4
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
lazy_loader==0.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
multidict==6.0.4
mutagen==1.46.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.3.0
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
py-cord==2.4.1
pycparser==2.21
pycryptodomex==3.17
Pygments==2.15.1
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytube==12.1.3
PyWavelets==1.4.1
pywin32==306
pyzmq==25.0.2
redis==4.5.4
requests==2.28.2
scikit-image==0.20.0
scipy==1.10.1
six==1.16.0
soupsieve==2.4.1
spotipy==2.23.0
stack-data==0.6.2
tifffile==2023.4.12
tornado==6.3.1
traitlets==5.9.0
urllib3==1.26.15
wcwidth==0.2.6
websockets==11.0.2
yarl==1.9.1
yt-dlp==2023.7.6
